+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Allow requests to portable inboxes/outboxes signed with keys identified by 'ap' URLs.
+- Added `webfinger` command.
 
 ### Changed